feat: export defineParameterType from Global.js
